Extend override report to include expired files
This extends the existing override_audit_override_report command to
take two boolean arguments (both of which default to False).

The first indicates that while generating the report we should also
get the list of expired overrides and mark up each override file and
package with text to indicate the expired state.

The second indicates that the report should contain only expired files
and nothing else. This is meant to be used in the automated checks to
cut down on noise when reports are generated by telling you
specifically what overrides are currently out of date.

In order to support this new functionality, the syntax for the
override report has been extended to mark up the new text for expired
files as well as to flag them with an additional scope qualifier so
that they can be targetted by context commands later.

Additionally, the context command for refreshing an existing report
was updated to support the new variants of the command.

With the new arguments, new command options have been added to the
command palette and main menu.