Properly filter expired only override report
This redacts any packages from the expired only override report if
they are not expired complete overrides and don't have any expired
simple overrides.

Previously a package would still show up in the expired only report if
it wasn't an expired complete override because the code only checked
the list of expired overrides.