Always display package status when decorated
This modifies the function responsible for decorating a package name
to include its shipped/installed/unpacked status so that it always
appends the "complete override" and "expired" status items when it is
appropriate, instead of optionally doing it via a paramter.

All places that use this method do (or currently will) always set both
values to true, so this just makes the code cleaner.

We can revisit if some new report may decide not to do this or
configure it away or some such.