Fix the syntaxes (yet again)
While looking into a completely unrelated problem I discovered that
the override list and diff views had problems properly matching the
name of a package which has spaces in it. That's something that I'm
pretty sure used to work before I modified things to include the
expired flag.

This change (hopefully) fixes that problem for good now.