Fix diffing of binary files
An issue which doesn't often crop up (but could, since it did to me)
is that it's probably not a good idea to perform a diff on a file that
is binary, such as a PNG file that is used in a package.

Previously such a file would cause the diff report to tell you to look
in the console for an error (which ironically showed no error), but
the ultimate cause of this is that binary files are generally not
UTF-8 compliant.

This change includes the ability for the override diff functionality
to detect when a file is binary and return a specific custom result
instead of trying to perform the actual diff in that case.

This is currently set up via the setting binary_file_patterns, which
is a standard Sublime Text setting.

If such a setting exists in the OverrideAudit.sublime-settings file,
it will be used as the binary pattern source; if not, the version from
the user preferences is used instead.

This allows users to customize the setting specifically for override
audit only if they need to, without having to have that setting be in
effect for all of Sublime Text (which uses it for other purposes), as
some people use it as a workaround to exclude files from indexing, for
example.