Rename PackageInfo member variables for clarity
For some refactoring that is currently in progress, some of the member
variables in the PackageInfo class are being given new names that
better reflect what they're for.

These are a separate commit because A) I like to commit at the drop of
a hat and B) these changes along with the upcoming changes are
breaking my head on the diff.