Add abilty to find zipinfo of override file
This adds to the PackageInfo class the ability to find the zipinfo
structure that represents an override file (either simple or complex)
given the name of the override file itself.

This works by having the package keep a cached list of all of the
zipinfo structures that represent its members, with an optional cached
dict version for quick lookup.

In use, when you ask for the zipinfo for an override, the dictionary
is used to directly look up the zipinfo for the given override, which
provides not only its name but also its modification time.

If the lookup fails and the system is currently case insensitive for
filenames, a secondary scan is done by checking for the same file in
a case insensitive manner.

This allows for the lookup of a file on e.g. Windows by a name given
by the user that may be of the wrong case and have the override still
be looked up.

This will be used for some upcoming bug fixes and tweaks.