Use new API to make expired overrides work better
Using the new API method, the code for getting the list of expired
overrides now checks inside of the package file that is providing the
source file in order to get the modification time of the archive
member for comparison to see if the local unpacked override file is
out of date or not.

This allows for things like updating the sublime-package file by
adding an updated copy of a file without making everything else from
the same package appear expired. It remains to be seen how the Sublime
devs handle this when packaging a new build, but the theory is sound
either way.

The new code falls back to using the source sublime-package timestamp
if it cannot find the appropriate entry in the zip file, which should
never happen. However, those are famous last words I don't want to
live to regret.

It seems better to provide too many warnings about an expired override
when something untoward happens than too few; at least with too many
you can double check everything if needed whereas with too few you
could blithely assume everything is OK when it in fact is not.