Add info on binary_file_patterns setting
This adds a stub entry with documentation comment to the settings file
to indicate that it is possible to set a custom list of files to be
treated as binary (and thus not diffed). This is mentioned in the
README but it's probably best to include in the settings as well,
since many people (myself included) may tend to look there first.