Clean up packages.py
This cleans up a bit of code in this particular file. Largely this is
to try and fit coding conventions as much as possible, but some things
such as documentation comments for newly added API functions were
missing and some older comments needed to be cleaned up a bit.

As of now the only things falling out of flake8 are errors regarding
lines being too long (mostly 80 instead of 79) and complaints about
my comment separators.

The first of those things I'm not going to address because the
readabilty of the code is higher with more verbose names and such
lines are few and far between.

The latter I'm not going to address because complaining about a
comment starting with "##" and not "# " seems like the height of "I
can't read your code because there are too many hash marks" style
arrogance.