Make package names case insensitive on some OS's
OverrideAudit uses the name of sublime-package files and directories
inside of the Packages\ directory to determine the names of packages.

Although this works fine on Linux as originally written, it can have
problems on Windows and MacOS, where the file system is not case
sensitive.

In particular, on those operating systems, Default.sublime-package
and Packages\DEFAULT are both considered to be the same package, and
sublime treats them as such. On Linux this is treated as two distinct
packages.

Although it's possible that nobody is doing such a thing (and it could
have disastrous consequences since Linux treats such things as
identical) we can't rule it out, and OverrideAudit strives to treat
packages the same way that Sublime itself does.

To this end, on systems which have case insensitive file systems, the
PackageList class is now case insensitive when it comes to package
names.

The names of packages is set to a "de facto" known state based on the
first time it is encountered during the package scan. Since we scan
first for shipped packages, then installed, then unpacked, this  means
that packages will (for the most part) be known by their "correct"
name.

Package names that are given to API functions that look up a package
or check if it exists will be case insensitive and return the correct
PackageInfo instance regardless of the case used to make the request
in these cases.

Internally this is done via a proxy dictionary which maps lowercase
versions of the package names to their "correct" names.

All code to access e.g. a package file or override still tries to use
the name as seen in the PackageInfo class, which should be the actual
path for case sensitive file systems and work regardless of case on
case insensitive systems, so this change should be wholly transparent.