Add paramter to skip override report when empty
This adds an optional parameter ignore_empty to the existing command
for generating an override report (either version) which causes the
report to do nothing if the list of overrides is empty. When not
specified, the default value of False is applied.

When this parameter is set to True and a report would otherwise be
empty, instead of opening up a view to tell you that there is nothing
interesting, it will instead display a message briefly in the status
bar and also into the console.