Allow for proper module reloading/updating
This tweaks up the main override_audit.py file so that whenever it is
loaded, it detects if its submodules (i.e. the contents of the lib
folder) have been loaded previously, and if so it manually reloads
them, as Sublime does not do this on its own.

Besides being handy for development purposes as simply saving the top
level file ensures that everything reloads, this is also needed to
allow for proper interaction with the ignored packages setting of
Sublime Text (and thus for package upgrades via PackageControl).

This aspect should be shifted out to an external top level module file
if we end up with more than a single top level python file, but that
seems redundant for the time being (and is not as handy for
development purposes).