Include an initial spinner class
This is a simple utility class which takes a window, thread and
caption, and then outputs an animated spinner in the status bar while
the thread is still running.

This is to be used while all of our threaded commands are churning in
the background so that the user is aware that something is going on.

The spinner will be constrained to the provided window (in which the
eventual output of the executing command will be displayed) but the
spinner "follows" the active view if that changes while it is running,
as the content of the status line is per-view.