Begin threading initiative
Henceforth we commence with the grand scheme to make our commands
thread based (or at least those for which it makes the most sense) so
that the time taken to e.g. crawl the package directory for
information does not hang the UI.

I'm trying to abstract away the bulk of the redundant logic for firing
up the thread, setting up the spinner, and generating the content in
the main thread away into a helper class so that the actual reports
can be clearer.

This may require a bit of tweaking in the future as we implement the
remaining commands.