Abstract away package name modifications
This is actually something I meant to do a ways back, but now seems as
good a time as any since the removal of other redundant code in
reports is going to make the code tighter.

Basically this shifts the code that decorates [Ignored] and
<Dependency> package names for display into a single place so that the
horror of that stupid chain only has to happen once.