Partially thread the single override diff method
This adds in partial threading for the command that allows you to diff
a single override.

Currently only the collection of the package list is threaded, by way
of a new helper class that will be used for some other commands. The
thread collects the package info (and optionally also collects the
information on overrides if the caller thinks it might need that
right away)

The interaction for selecting the package and override remains in the
main thread (and shall remain there as that's the place for it), but
the diff is currently also still done in the main thread.

That aspect will be coming soon, as it is used for the context
commands as well.