Fix the spinner pooping in the status line
Due to a flub in the spinner code, it was possible in some cases (e.g.
multiple spinners at once) to leave itself in the status line.

This was due to not correctly clearing away a view properly when the
spinner noticed that it changed from the last call. In this case when
one operation completed (generally changing the active view to the
new report), things would get left in an inconsistent state.

Props to @keith-hall for quickly spotting the issue during review.