Refactor thread processing for the package list
This is a slight tweak to things to have a core base class that
represents a worker thread. The class takes care of showing the
spinner and running processing, checking periodically to see if things
are done or not.

Once completed, the thread can optionally also invoke a callback in
the main Sublime thread with a single parameter that is the thread
instance, allowing results to be collected.

This new base class collects keyword arguments so that subclasses can
take whatever arguments they require to work.

The object for collecting the package list in the background has been
reworked to use this new mechanism.