Thread single override diffing
Since all commands that perform a diff of a single override always
invoke the global function to calculate and display the diff, this has
been abstracted away by using a background worker to calculate the
diff directly from that function.

This means that everything that performs a diff automatically gets
the diff performed in a thread "for free".

The function was renamed to better indicate what's happening when you
invoke it.