Make event handling async for loading/saving
OverrideAudit implements an event listener to know when a file is
being loaded or saved so that it can check to see if it's possibly an
override or not, allowing integration with the context menu to allow
for easy diffing of overrides opened manually (say via
PackageResourceViewer).

This change swaps the on_load and on_post_save for the async versions
of those events to minimize any slowdown. The code in them quickly
short circuits away if the file provided can't possibly be an override
(i.e. is it not rooted somewhere within the unpacked packages
directory), but for such files it needs to do a scan to see if a
package exists or not, which does not take zero time.