Implement automatic reports on package unignore
This adds in the ability for the plugin to notice when a package is
being removed from the global list of ignored packages in sublime so
that it can generate an automatic report on expired overrides.

PackageControl adds packages to this list while it is working with
them (installing, upgrading, etc), so this will cause the report to be
generated after package operations complete.

This also triggers just for any package un-ignore under the theory
that if you are turning on a package that has been turned off, you may
be exposing yourself to issues with an expired override.

The code implements a delay between when a package being removed from
the ignored list happens and when the report triggers. Any extra
removals in this delay will reset the delay, so that if there are
multiple packages being upgraded, the report only generates after all
of the upgrades are complete.

Additionally, if all unignored packages are added back to the ignored
list prior to the report being generated, the report is canceled.

The default delay is 5 minutes, but the delay is user configurable and
can be set to 0 to disable this feature in its entirety.