Update AutoReportTrigger to trigger on startup
More specifically, we store a file in the User package which contains
the version of Sublime as of the last time we did an automatic report
as well as a boolean indication of whether a report should be forced
to display at startup.

When Sublime loads, we load the contents of this file if it exists.
When the stored version of Sublime is not the same as the current
version or the boolean is True, we trigger an automatic report at
startup. Both of these will always be true when the file is missing.

The status is saved whenever we decide to trigger a report, when we
decide to cancel a triggered report, and when a report that was
triggered actually happens (after the user specified delay).

These changes allow us to automatically generate a report at startup
every time the installed version of Sublime changes or when sublime is
restarted while a report is pending (perhaps due to a package
upgrade).