Update README for the new automatic report code
This includes a section on what the automated reports are, when they
run, and what configuration option exists to control when they happen.

As of right now, only the automatic reports on package un-ignore can
be turned off; reports when the version of Sublime Text updates are
currently always turned on.