Add logging method
Now that we're in the final stretch of polish before an initial
release, this adds in a _log method which will send output to the
console and optionally also the status bar of the currently active
window, with the console text prefixed with the name of the package so
that it is more clear what is generating it.

As cleanup/testing progresses, more logging will be added in an
attempt to ensure that it is always possible for a user to determine
what is happening or what might have gone wrong.