Change automatic reports on package upgrade
Initially the code would schedule an automatic report for a
configurable amount of time after a package was removed from the list
of ignored packages, so that multiple removals (e.g. due to many
packages being upgraded at once) would not cause multiple reports.

However it appears that PackageControl removes all packages it is
upgrading all at once at the end, so this delay is not necessary and
may actually get in the way.

Things have been modified now so that the setting is a boolean toggle
instead of a time delay, allowing you to still turn the automatic
reports in this case off if desired.