Clean up syntaxes and add tests
This is some minor changes to the syntaxes, slightly changing some
scopes and file names and (most importantly) adding in some syntax
tests to assure that things are working as expected.

The full relative filenames of the syntaxes was quite large and also
contained the text OverrideAudit in the filename, which is redundant
since it is also the name of the package.

This change changes the names of the files to OA-something from
OverrideAudit-something, as well as slightly changing the names and
scopes so that lists are reports (as they are referred to everywhere
else).

Internally in the code a helper method has been included which gets
the full name of the file from a small part, which makes things a
little cleaner to look at and also shortens the lines up enough to
not break the 80 column rule.

The names of the files still have OA- in the name because the name of
the syntax file is the name of the settings files for syntax specific
purposes, and users may wish to have specific settings in place for
our reports (e.g. to turn off wrapping in diffs).

As a side note to this, the syntax for bulk diffs has been extended to
properly highlight a dependency even though a dependency cannot have
any overrides due to it only being an unpacked package.

This is because it's possible to use the context menu from the package
list to select a dependency, and the diff works as expected in this
case (showing no overrides as it should) so the dependency name should
be colored properly as well.