Refactor settings into a single function call
In order to help with code clarify, there is now a helper function
named _oa_setting() which takes an OverrideAudit setting name and
returns back the currently configured value for that setting.

This is done using a cached settings object, so all code for getting a
setting can just give the setting name and not have to make sure to
load the settings.

Additionally, the function has a list of the default values for all
known settings and will return that if a setting is missing from the
configuration object.

This removes the redundancy of having to always specify the default
when fetching from the settings object, which is potentially error
prone if/when the default value of the setting is changed.