Refactor packages.py settings for binary
The code in the lib folder is meant to be reusable and have nothing
OverrideAudit specific in it.

To that end it was a bit of an initial crufty hack to have the diff
code pull settings from OverrideAudit.sublime-settings for the list of
binary patterns, and then from the user preferences if it was not set.

Now the PackageInfo class assumes that the list of binary files is the
Sublime default (from global preferences) and there is a method to
change it to a different list.

Accordingly the code in override_audit.py checks if there is a
specific setting for binary_file_patterns in the OverrideAudit
settings file and if so, it applies it prior to a diff.