Merge package decoration into a single function
In order to clean things up a bit, merge together the function for
decorating a package name with [] or <> to denote its status with the
method that also applies a prefix and suffix to indicate the type of
the package and if it is expired or has a complete override.

The single method takes a parameter that will short circuit and return
back just the decorated name if desired.

This commit also covers a slight change to the default configuration
information in the code, in which the default value for one of the
settings was accidentally left in the wrong state during testing.