Enhance Preferences integration
This change modifies how OverrideAudit integrates with the Preferences
system in Sublime text, including offering a default empty settings
file when attempting to edit settings and including the preferences
options in the command palette (which in my opinion not enough
packages do).

Additionally, this change includes two simple proxy commands for the
default open_file and edit_settings commands for use in the
preferences menu.

open_file is present in all versions of Sublime that we support, but
edit_settings was added in build 3124. When this happened Sublime
switched from offering two menu items for everything (Default and
User) to a single entry that opens in a window with the default and
user settings side by side.

The proxy commands just invoke the underlying command, but they hide
and show themselves from the menu and command palette based on the
current build number of Sublime, so that in newer versions there is a
single entry while for older versions (which do not support the
command) there are two.