Clean up view settings for overrides and reports
OverrideAudit uses custom view settings to indicate if a view contains
either the contents of an override, the diff of an override, or a
report of some sort. In each case the presence of the settings
indicate that the view is special, while the values of the settings
indicate the underlying data.

This commit clarifies this code, primarily by way of a utility class
that takes care of the details of getting, setting and testing a group
of settings all at once and by removing a redundant global function
now that the report code has been mostly consolidated into a single
threaded base class.

Although the number of lines in the file doesn't change much, the
readability (to my eyes) goes up a bit because everything looks
cleaner and it's more obvious from looking that the settings are
special.