Enhanced on_query_context to support reports
This enhances our key context handling, adding a context for report
views and supporting the equal and not equal operators where
previously we assumed that the user always wanted to test the context
as equal to true.

All of the default keymap files now contain an example of using this
new context to be able to refresh the current report as an example of
the context and how it works.

This new binding is commented out of all keymaps (with instructions on
how to use it) because of lack of a good default.

It seems like the browser refresh keys drilled into user memory would
be a good candidate. On Windows/OSX browsers tend to use F5 and Ctrl+R
to signal a refresh, while on MacOS Option+R is used.

However, Ctrl+R and Option+R are mapped by default to open Goto
Symbol, which our reports support to allow quick navigation among the
packages and overrides.

F5 might be a candidate except that on MacOS this is not really used
as a refresh key, and Sublime maps it by default to the sort lines
command.