Officially uncomment the report refresh bindings
After some extra thought and discussion, this change uncomments the
key bindings for refreshing a report, using the F5 key across the
board.

Although this key is currently bound to sorting lines under MacOS, it
has been pointed out that since reports are read only anyway, that's
not too much of an issue.

Conceivably it may bother pure MacOS users that this key is the one
used for refreshing when they're not used to it, but I think the
benefits outweigh that, since such a user can always change the
binding anyway and the default isn't going to hurt anything.