Display console errors in packages.py
This is a modification to the methods in PackageInfo that return the
contents of a file (either packed or unpacked) so that when they
return None to indicate there was an error, a message is sent to the
console to indicate the probable reason.

This allows for someone to figure out why things have gone wrong,
which is important e.g. for when you manually open a file that might
be an override but actually is not.