Enhance _log to also show a message box
This is a simple enhancement to the _log() function that we use to
display error messages to the console and optionally also to the
status bar so that it can also optionally display a message dialog to
the user as well.

This will be used to draw user attention to problems when they take an
explicit action, as the text in the status line is not always a
greatest attention draw.