Display errors on single override diff
Using the new _log() functionality and the error messages that are now
displayed by the methods in PackageInfo() for getting file contents,
display an error if a diff fails.

Due to the threaded nature of the diff operation, this currently
displays just an error message telling you what you were trying to
diff when the error happened, directing your attention to the console
to get information on what went wrong.

In a future version this will be able to display that information and
also propagate the information back to the view that issued the diff
command from the context menu or command palette so that it knows it
should remove the settings that mark it as an override.