Enhance the Bulk Diff command
This adds a few enhancements to the bulk diff command, in both the
case of being told to bulk diff a single package as well as when used
to bulk diff all packages.

First, when you attempt to bulk diff a package that doesn't exist, the
error message is displayed in the status line and in a dialog popup;
previously, this was only logged into the console.

Second, when there are no packages with overrides, the report says
that this is the case instead of saying "Bulk diff of overrides in 0
packages" which, while accurate, is not very user friendly.

Lastly, if the command is used to bulk diff a package that cannot have
any overrides (e.g. from the Package Report), instead of saying "[No
simple overrides found]" the message says that there are no overrides
and says why there are none ("no unpacked files" or "no sublime-
package file"). This makes the message more informative.