Enhance Diff Single Override command
This adds in a couple of simple enhancements to the command. The
existing _log() about there being no packages with overrides now also
sends its output to the status bar and to an error dialog, so that
when selecting the command manually something happens even if it's
just an error message.

This ensures that there is always a visible effect when the command is
executed, even if there's nothing to diff.

Additionally some extra error messages were included for cases in
which the command gets a package and override file directly when one
or both does not exist.