Fix thread latency and a thread object leak
As originally implemented, the background threading code was using a
repeated sublime.set_timeout() to check and see if the thread was done
running yet or not.

After that point the code was redesigned so that the object
constructor takes a callback, which would be invoked when the thread
was finished processing. However, the same mechanism for checking to
see if the thread was done was left in.

Although this works it adds extra latency to commands, which would
poll first 10 milliseconds after starting, and then every 100
milliseconds after that.

Now when the thread finishes running it will immediately use
set_timeout() to schedule the callback in the main thread, which cuts
the latency quite a bit.

At the same time there was a thread object leak in the existing model
in that the callback was stored as a property of the thread object,
and took the thread object as a parameter. This locked the thread
forever, stopping it from being collected.

To get around this the code first clears away the property and then
schedules the callback so that a circular reference back to itself is
not made.