Clarify what the "diff_unchanged" setting does
This clarifies in the default settings file what this is for, which is
specifically what to do when using the "Diff Single Override" command.

In particular, when performing a diff via the context menu, command
palette or key binding the command always assumes that the value is
set to "diff" regardless of user setting, so that there is always a
confirmation that what you tried actually did something.

In addition to this, the status message text that gets displayed when
this is set to "ignore" in the single command case has been extended
to indicate exactly what package and override the diff was attempted,
and it is now also sent to the console as well as to the status line.