Fix refreshing reports when reuse_views is false
This is a use case that I didn't initially test, but refreshing an
existing report would not perform as expected if the "reuse_views"
setting was set to False (it is True by default).

Under this setting every report command always creates a new view,
even when an existing one already exists. You could use this as a
before and after, for example.

In this case, refreshing a report would always refresh the first
report with the correct title found in the window even if that wasn't
the one the user wanted refreshed.

The code for finding a view to reuse for a command now takes an
optional parameter that indicates what the current view is, and that
view is tested to see if it is the correct one before the search for
the existing view kicks off.

In addition the report generation command now makes sure that the view
that is the target of the command is the active view, which is
captured at the time the generation thread is kicked off.

This ensures that in this situation it is possible to refresh a
specific report when there is more than one, although it does mean
that the focused view switches to the one being refreshed.

For the most part I think this is OK; where it could potentially be
jarring is when the command is targeted via a context menu command
on a tab in another view group or window, where the report may not
currently have the focus.