Small tweaks to the README file
In preparation for release, this cleans up the README file slightly,
adding in some last minute tweaks and removing the information about
provisionally installing via Package Control (although a request has
not yet been made as final testing is not yet completed).

In addition the License information has been redundantly put in a
LICENSE file so that people that ignore the very obvious instruction
implicit in the name of the README might still notice that there is a
license and read it.