Fix a case problem with override diffing
In the Sublime text internals, the core doesn't take the case of
filenames into account for package folders and package contents if the
file system itself is not case sensitive, and OverrideAudit takes
pains to make sure that it operates the same way.

In general that means that whenever working with overrides we make
sure to use the files as contained inside of the sublime-package
files themselves, which are considered to be the "gospel". If the user
file is not the same case as stored in the package, everything still
works because the file system doesn't care.

There was an issue with this in the code that does checks when a file
is manually opened and it has the wrong case. The context menus get
applied if the file might be an override, but when you try to diff the
name of the file is looked up in the sublime-package file using the
local name.

Since zip files are case sensitive, this would fail, so having an
override named Default\EXec.py and manually opening it would cause
a diff failure because no file by that name is in the zip.

The code for extracting the contents of a packaged file now has extra
logic that applies for systems with case insensitive file systems that
will do a case insensitive search if needed to try and find the file,
which allows the contents to be loaded.