Compatibility fix for status_message()
The code in the _log() method was invoking the status_message() method
from the currently active window, but this was introduced in Sublime
3110 and prior to that you had to use sublime.status_message()
instead.

Some investigation seems to indicate that at least in build 3126,
sublime.status_message() and sublime.active_window().status_message()
do the same thing; display the status message only in the currently
active window.

Back in Sublime build 3092 sublime.active_window().status_message()
does not exist, but sublime.status_message() displays the message in
the status bar of all windows.

In order to be compatible with older versions of Sublime and since
both do the same thing, the _log() code now uses the sublime module
version when requested.