Compatibility fix for _thr_diff_override
In order to make OverrideAudit compatible with version 3092 of Sublime
this changes the syntax assigned to single file diff views in
_thr_diff_override() to use the tmLanguage version instead of the
sublime-syntax.

Although Sublime build 3092 supports the sublime-syntax format of
files, at the time of its release the shipped format for the Diff
syntax was still tmLanguage and not sublime-syntax, which was changed
in a later version.

Newer versions of Sublime support both, and in particular when a
syntax is requested using tmLanguage and there is not one but there is
an  associated sublime-syntax, Sublime will correctly use it.

This change ensures that older versions from prior to the Diff syntax
swap will still work as expected.