Fix #14 - Mixed path separators in subfolder diffs
These changes standardize all of the path displays that occur from
within OverrideAudit so that they always use the unix style forward
slash.

This is how overrides were previously always displayed due to that
being the native way that sublime-package files store the path
fragment. The real issue here was in prefixing the name of the package
and the package location using a path join, which would use the native
platform path separator.

Although it would make sense to standardize all display paths to the
platform path separator, as has been pointed out by @FichteFoll,
Sublime uses this path style for files relative to the Data directory
(which all override and override source files are), so this way seems
more in line  with how Sublime would portray the same paths.