Fix #13 enhanced native override integration
Using some code that was implemented prior to Version 1.0 code freeze
and taking advantage of the event listener using async events, this
enhances the checks done at file load and save time.

Previously, the context menu was enabled for any file that appeared to
be within the Packages folder for a package that had a sublime-package
file present.

This could present the context menu for editing and diffing when it
should not because a file in the package directory that's not inside
the package isn't an override.

This now correctly checks to see if the file being opened in this case
is actually contained in the sublime-package file being used by
sublime or not, which stops this from happening.

This was originally in the 1.1 milestone but I've pulled it out of
there because the next release won't be a point release, just a
maintenance release.