Rename Diff tab and command to be more meaningful
This is a small cosmetic change that alters the tab text that is used
for a single diff file to be "Diff of filename" instead of "Override
of filename", which is less meaningful.

As a result of this the command text for the command to swap between
the diff and edit in the command palette has been similarly tweaked
for clarity.

This also pointed out to me that I completely forgot to document the
diff_unchanged setting, which is now fixed.