Include GitHub contribution guidelines
This includes a contributing document that outlines how best to report
a bug, request a feature or get in touch with the developer. Also
included is an issue template to help guide issues and make them more
consistent.

There is as yet no template for pull requests based entirely on my
having no idea whatsoever what the best practices for that will be
until/when it starts happening.