Include a save_on_diff option to the diff command
This includes a new configuration option save_on_diff which, if set to
true, will save any unsaved changes in an edited override file to disk
before performing the diff.

Due to the principal of least surprise, the option defaults to false
and will not persist changes to a modified buffer if the underlying
file has been removed. This ensures that the user never unwittingly
persists changes or resurrects a deleted file