Get ready for point release 1.0.1
This rolls up the initial tweaks and additions made to OverrideAudit
during the first week of release.

Along with some visual changes, work flow is improved a bit with the
addition of the ability to automatically save changes in an override
edit view before switching to/refreshing a diff, making it easier to
see the effects of your changes.