Fix #24 support overrides with a leading period
This issue was caused by the syntax in the diff report requiring the
name of an override to be a word character, followed by anything else
all the way to the end of the line.

This has been fixed (and new tests added to cover the situation) by
allowing any character for the entirety of the part of the line that
contains the override filename to be a part of the name.

This could still conceivably cause problems with people having other
weird filenames, such as a leading space or something with the
explicit text "[X] " on the front, but you can only protect people so
much.